🎨 : SearchBar 파일 이동 및 Header 기능별로 분류(UserStatus)
