+++ src/pages/SearchList.tsx
@@ -1,0 +1,3 @@
+export default function SearchList() {
+  return <div>SearchList</div>;
+}
